TodoState.name を使う
