+++ app/lib/model/Todo.scala
@@ -1,6 +1,7 @@
 package lib.model
 
 import ixias.model._
+import ixias.util.EnumStatus
 import lib.model.Todo.Id
 
 import java.time.LocalDateTime
@@ -10,7 +11,7 @@ case class Todo(
     categoryId: TodoCategory.Id,
     title:      String,
     body:       String,
-    state:      TodoState,
+    state:      Todo.TodoState,
     updatedAt:  LocalDateTime = NOW,
     createdAt:  LocalDateTime = NOW
 ) extends EntityModel[Id]
@@ -18,4 +19,13 @@ case class Todo(
 object Todo {
   val Id = the[Identity[Id]]
   type Id = Long @@ Todo
+
+  sealed abstract class TodoState(val code: Short, val name: String)
+      extends EnumStatus
+
+  object TodoState extends EnumStatus.Of[TodoState] {
+    case object Todo       extends TodoState(code = 0, name = "TODO(着手前)")
+    case object InProgress extends TodoState(code = 1, name = "進行中")
+    case object Done       extends TodoState(code = 2, name = "完了")
+  }
 }
